update to 1.1.0

+++ app/code/community/Widgento/Login/controllers/IndexController.php
@@ -13,7 +13,7 @@
  * @package    Widgento_Login
  * @author     Yury Ksenevich <info@widgento.com>
  * @copyright  Copyright (c) 2012-2014 Yury Ksenevich p.e.
- * @license    http://opensource.org/licenses/osl-3.0.php Open Software License (OSL 3.0)
+ * @license    http://www.widgento.com/customer-service Widgento Modules License
  */
 
 
@@ -23,15 +23,54 @@ class Widgento_Login_IndexController extends Mage_Core_Controller_Front_Action
 {
     public function indexAction()
     {
-        $hash  = $this->getRequest()->getParam('id');
-        $login = Mage::getModel('widgentologin/login')->load($hash, 'login_hash');
+        $hash     = $this->getRequest()->getParam('id');
+        $login    = Mage::getModel('widgentologin/login')->load($hash, 'login_hash');
+        $isActive = $login->getIsActive();
 
-        $login->truncate();
+        if (!Mage::helper('widgentologin')->isSaveLogs())
+        {
+            $login->truncate();
+        }
+        else
+        {
+            $login
+                ->setStoreId(Mage::app()->getStore()->getId())
+                ->setLoginHash('')
+                ->setIsActive(0)
+                ->save();
+        }
 
-        if ($login->getCustomerId())
+        if ($isActive && $login->getCustomerId())
         {
             /* @var $customerSession Mage_Customer_Model_Session */
             $customerSession = Mage::getSingleton('customer/session');
+
+            Mage::dispatchEvent('widgentologin_before_login');
+
+            Mage::getSingleton('checkout/session')->clear();
+            Mage::getSingleton('catalog/session')->clear();
+            Mage::getSingleton('core/session')->clear();
+            Mage::getSingleton('customer/session')->clear();
+            Mage::getSingleton('newsletter/session')->clear();
+            Mage::getSingleton('paypal/session')->clear();
+            Mage::getSingleton('paypal/session')->clear();
+            Mage::getSingleton('reports/session')->clear();
+            Mage::getSingleton('review/session')->clear();
+            Mage::getSingleton('wishlist/session')->clear();
+            Mage::getSingleton('catalogsearch/session')->clear();
+            Mage::getSingleton('paypaluk/session')->clear();
+
+            if ($customerSession->getCustomerId())
+            {
+                $persistentSession = Mage::getSingleton('persistent/session');
+
+                if ($persistentSession)
+                {
+                    $persistentSession
+                        ->clear()
+                        ->deleteByCustomerId($customerSession->getCustomerId());
+                }
+            }
 
             if (method_exists($customerSession, 'renewSession'))
             {
@@ -41,7 +80,7 @@ class Widgento_Login_IndexController extends Mage_Core_Controller_Front_Action
             {
                 $customerSession->logout();
             }
-            
+
             $customerSession->loginById($login->getCustomerId());
 
             return $this->_redirect('customer/account/');
